requirements.txt
flask
flask-sqlalchemy
flask-login
flask-cors
playwright
cryptography
python-dotenv
apscheduler
requests
beautifulsoup4